Дипломная

+++ src/main/java/org/project/ghost_forum/service/CommentService.java
@@ -2,19 +2,18 @@ package org.project.ghost_forum.service;
 
 import lombok.RequiredArgsConstructor;
 import org.project.ghost_forum.dto.CommentDto;
+import org.project.ghost_forum.dto.UserDto;
 import org.project.ghost_forum.entity.Comment;
-import org.project.ghost_forum.entity.Post;
 import org.project.ghost_forum.entity.User;
+import org.project.ghost_forum.exception.CreatingException;
 import org.project.ghost_forum.mapper.CommentMapper;
 import org.project.ghost_forum.mapper.UserMapper;
 import org.project.ghost_forum.repository.CommentRepository;
-import org.springframework.context.annotation.Lazy;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -25,17 +24,12 @@ public class CommentService {
     private final CommentRepository repository;
     private final CommentMapper mapper;
     private final UserService userService;
+    private final UserMapper userMapper;
 
-
-    public Set<Comment> getComments(List<UUID> commentIds){
-        return commentIds.stream()
-                .map(repository::findById)
-                .map(Optional::get)
-                .collect(Collectors.toSet());
-    }
 
     @Transactional
-    public CommentDto newComment(CommentDto commentDto){
+    public CommentDto newComment(CommentDto commentDto) {
+        if (commentDto.getBody().isEmpty()) throw new CreatingException("Комментарий не может быть пустым.");
 
         commentDto.setCreationTime(LocalDateTime.now());
         commentDto.setUserId(userService.getCurrent().getId());
@@ -47,28 +41,33 @@ public class CommentService {
     }
 
     @Transactional
-    public CommentDto editComment(CommentDto commentDto){
-        Comment comment = repository.findById(commentDto.getId()).orElseThrow();
+    public CommentDto editComment(CommentDto commentDto) {
+        if (commentDto.getBody().isEmpty()) throw new CreatingException("Комментарий не может быть пустым.");
 
-        Comment savedComment = repository.save(mapper.merge(commentDto, comment));
+        UserDto user = userService.getCurrent();
 
-        return mapper.toDto(savedComment);
-    }
+        if (user.getId().equals(commentDto.getUserId())) {
+            Comment comment = repository.findById(commentDto.getId()).orElseThrow();
 
-    public Comment getCommentById(UUID id){
-        return repository.findById(id).orElseThrow();
+            Comment savedComment = repository.save(mapper.merge(commentDto, comment));
+
+            return mapper.toDto(savedComment);
+        } else throw new AccessDeniedException("Ошибка! Нет прав на редактирование.");
+
     }
 
     @Transactional
-    public void deleteComment(UUID id){
-        repository.findById(id).map(comment -> {
+    public void deleteComment(UUID id) {
+        User user = userMapper.toEntity(userService.getCurrent());
+        Comment comment = repository.findById(id).orElse(null);
+
+        if (user.getId().equals(comment.getUser().getId()) || (userService.hasAdminAuthority(user.getId()))) {
             repository.delete(comment);
-            return null;
-        }).orElse(null);
+        } else throw new AccessDeniedException("Ошибка! У пользователя нет прав на удаление.");
     }
 
-    //Иначе ошибка неизвестного рода
-    private CommentDto transformToDto(Comment comment){
+    //Иначе ошибка
+    private CommentDto transformToDto(Comment comment) {
         return CommentDto.builder()
                 .id(comment.getId())
                 .postId(comment.getPost().getId())
@@ -79,9 +78,9 @@ public class CommentService {
                 .build();
     }
 
-    //Комменты к посту
-    public List<CommentDto> getCommentsToPost(UUID postId){
+    //Получаем комменты к посту
+    public Set<CommentDto> getCommentsToPost(UUID postId) {
         return repository.findAllByPost(postId).stream()
-                .map(this::transformToDto).toList();
+                .map(this::transformToDto).collect(Collectors.toSet());
     }
 }
